requirements.txt
arcade==2.3.8
asgiref==3.2.10
astroid==2.4.1
attrs==19.3.0
audioread==2.1.8
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
DevSqlite3==0.0.1
Django==3.0.7
djangorestframework==3.11.0
entrypoints==0.3
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipython-sql==0.4.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
json5==0.9.5
jsonschema==3.2.0
# jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
# jupyterlab==2.1.5
jupyterlab-server==1.2.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
librosa==0.7.1
llvmlite==0.30.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mpmath==1.1.0
# mysqlclient==2.0.1
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numba==0.46.0
numpy==1.18.1
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==6.2.2
prettytable==0.7.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
pycparser==2.19
pyglet==1.5.0
Pygments==2.6.1
pylint==2.5.2
# pyodbc==4.0.30
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytiled-parser==0.9.3
pytz==2020.1
# pywin32==228
# pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.24.0
resampy==0.2.2
scikit-learn==0.22
scipy==1.4.0
Send2Trash==1.5.0
six==1.14.0
SoundFile==0.10.3.post1
SQLAlchemy==1.3.18
sqlparse==0.3.1
sympy==1.6
tensorboard-logger==0.1.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
# torch==1.5.1
# torch-stft==0.1.4
# torchvision @ file:///D:/work/Pycharm/similar_image_generation/SRGAN-PyTorch/torchvision-0.6.1-cp37-cp37m-win_amd64.whl
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
wcwidth==0.2.4
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
gunicorn
